src/requirements.txt
gunicorn>=22.0, <23.0
uvicorn>=0.23, <0.24
fastapi>=0.111, <0.112
openai
python-dotenv
jinja2